Amazon app: Cannot interact with product page after tapping on product image https://bugs.webkit.org/show_bug.cgi?id=133559 <rdar://problem/16332749>
Reviewed by Oliver Hunt.

Source/WebCore:
As a workaround for issues in the Amazon app, mark the attribute navigator.geolocation as
replaceable so that we don't throw a JavaScript type error when assigning to an instance
attribute that shadows it. At the time of writing the first such assignment to is ignored.
See <https://bugs.webkit.org/show_bug.cgi?id=133648> for more details.

When loading a product page in the Amazon app it overwrites the readonly property geolocation
on an object that extends the Navigator object. In strict mode, this operation is not honored
and is treated as a JavaScript type error, which stops JavaScript execution. Therefore, the
product page does not behave as intended. In contrast, this operation would not be honored in
non-strict mode and would not be treated as a JavaScript type error.

Test: js/dom/shadow-navigator-geolocation-in-strict-mode-does-not-throw.html

* Modules/geolocation/NavigatorGeolocation.idl:

LayoutTests:
Add a test to ensure that assigning to an instance attribute in strict mode that shadows
navigator.geolocation doesn't cause a JavaScript type error.

* js/dom/script-tests/shadow-navigator-geolocation-in-strict-mode-does-not-throw.js: Added.
(createObjectWithPrototype.F):
(createObjectWithPrototype):
* js/dom/shadow-navigator-geolocation-in-strict-mode-does-not-throw-expected.txt: Added.
* js/dom/shadow-navigator-geolocation-in-strict-mode-does-not-throw.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@169710 268f45cc-cd09-0410-ab3c-d52691b4dbfc