Avoid creating a CFData when checking if a resource is file backed https://bugs.webkit.org/show_bug.cgi?id=133783
Reviewed by Andreas Kling.

Source/WebCore:
Export SharedBuffer::hasPlatformData().

No new tests because no functional changes.

* WebCore.exp.in:
* platform/SharedBuffer.h:

Source/WebKit2:
When a resource is file backed, we have it as a single CFDataRef in SharedBuffer. Add an
early return in tryGetShareableHandleFromSharedBuffer() so we don't end up creating a new
CFDataRef if we don't already have one in SharedBuffer. If we had to create a CFDataRef, the
it can't have been a file backed resource.

* NetworkProcess/mac/NetworkResourceLoaderMac.mm:
(WebKit::NetworkResourceLoader::tryGetShareableHandleFromSharedBuffer):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@169883 268f45cc-cd09-0410-ab3c-d52691b4dbfc