Page::findStringMatchingRanges() should take Vector<RefPtr<Range>> by reference instead of pointer https://bugs.webkit.org/show_bug.cgi?id=133677
Reviewed by Anders Carlsson.

Source/WebCore:
* WebCore.exp.in: Update the changed symbol.
* page/Page.cpp:
(WebCore::Page::findStringMatchingRanges): The method expects the matchRanges parameter to be non-null,
so it should take in that parameter by reference instead of pointer.
* page/Page.h:

Source/WebKit2:
* WebProcess/WebPage/FindController.cpp:
(WebKit::FindController::findString): Update the call to Page::findStringMatchingRanges() to pass
in a Vector<RefPtr<Range>> reference instead of a pointer.
(WebKit::FindController::findStringMatches): Ditto.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@170012 268f45cc-cd09-0410-ab3c-d52691b4dbfc