resolve-ChangeLogs fails when WebKit repository is a git submodule https://bugs.webkit.org/show_bug.cgi?id=134734
Reviewed by David Kilzer.

When using resolve-ChangeLogs in a WebKit git repository with a
non-default arrangement of GIT_DIR and GIT_WORK_TREE, incorrect
paths are computed and the script fails mysteriously.

We were naively computing GIT_WORK_TREE as being one level above
GIT_DIR, but this is not the case for submodules or a manually
configured GIT_WORK_TREE or GIT_DIR location.

* Scripts/VCSUtils.pm:
(gitTreeDirectory): Added. Find the equivalent of GIT_WORK_TREE.
(determineVCSRoot): Calculate the working tree root directly.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@170900 268f45cc-cd09-0410-ab3c-d52691b4dbfc