Pasting rich content does not perform a two-step paste https://bugs.webkit.org/show_bug.cgi?id=134795
Reviewed by Benjamin Poulain.

Source/WebCore:
UIKit requires writing something (anything!) to a particular pasteboard
during copy in order to flag the content as rich. When a paste occurs,
UIKit will read this flag and perform a two-step paste.

Test: platform/ios-sim/editing/pasteboard/two-step-paste.html

* platform/ios/PlatformPasteboardIOS.mm:
(WebCore::PlatformPasteboard::write):

LayoutTests:
Copy and paste some text, and make sure that the two step paste occurred.
The mechanism for detecting this is whether or not styles have changed appropriately.

* platform/ios-sim/editing/pasteboard/two-step-paste-expected.txt: Added.
* platform/ios-sim/editing/pasteboard/two-step-paste.html: Added.

Tools:
Two-step paste requires setting InputViewObeysDOMFocus on the UIWebBrowserView.

* DumpRenderTree/mac/DumpRenderTree.mm:
(createWebViewAndOffscreenWindow):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@170986 268f45cc-cd09-0410-ab3c-d52691b4dbfc