Rename selectorListContainsUncommonAttributeSelector() to selectorListContainsAttributeSelector() https://bugs.webkit.org/show_bug.cgi?id=134862
Reviewed by Sam Weinig.

Unlike containsUncommonAttributeSelector(), selectorListContainsUncommonAttributeSelector() does not
evaluate the attribute for "uncommon" types.

It would be possible to change the function instead to evaluate common attributes based
on the match type and the pseudo class type. Such change would be more risky
and we would get very little benefit from it, I leave that for later if that ever becomes useful.

* css/RuleSet.cpp:
(WebCore::selectorListContainsAttributeSelector):
(WebCore::containsUncommonAttributeSelector):
(WebCore::selectorListContainsUncommonAttributeSelector): Deleted.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@171043 268f45cc-cd09-0410-ab3c-d52691b4dbfc