Document::unregisterNodeListforInvalidation() and Document::unregisterCollection() have incorrect assertions https://bugs.webkit.org/show_bug.cgi?id=134869
Reviewed by Darin Adler.

Both methods should assert that the relevant HashMap is either empty if invalidation originates
from Document::invalidateNodeListAndCollectionCaches() or acutally contains the element that is
being invalidated. In the first case the HashMap is empty because its entries were moved out in
the Document::invalidateNodeListAndCollectionCaches().

This was exposed by r170995 (later rolled out in r170999) which introduced move constructor and
move assignment operators for HashTable. The assertions in the titular methods won't be passing
until r170995 relands.

* dom/Document.cpp:
(WebCore::Document::unregisterNodeListForInvalidation):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@171261 268f45cc-cd09-0410-ab3c-d52691b4dbfc