ASSERTION FAILED: generatingElement() in WebCore::RenderNamedFlowFragment::regionOversetState https://bugs.webkit.org/show_bug.cgi?id=135153
Reviewed by David Hyatt.

Source/WebCore:
Even though the CSSRegions spec defines the behaviour of a multicolumn region,
we currently do not support this functionality. This patch ensures that a multicolumn
element does not become a region. In the future, when we will implement the multicolumn
as region functionality, http://dev.w3.org/csswg/css-regions/#multi-column-regions, we
will remove this restriction.

Test: fast/regions/multicol-as-region-prevented.html

* rendering/RenderBlockFlow.cpp:
(WebCore::RenderBlockFlow::createRenderNamedFlowFragmentIfNeeded):

LayoutTests:
* fast/regions/multicol-as-region-prevented-expected.html: Added.
* fast/regions/multicol-as-region-prevented.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@171476 268f45cc-cd09-0410-ab3c-d52691b4dbfc