CSS viewport units use the wrong viewport size in WebKit1 https://bugs.webkit.org/show_bug.cgi?id=135254 <rdar://problem/17781423>
Reviewed by Tim Horton.

Source/WebCore:

Test: fast/css/viewport-units-dynamic.html

In WebKit1 on iOS, we want to resolve viewport units against the visible
viewport, not the legacy WK1 notion of the "viewport" which is the entire document.

Fixes rendering of medium.com articles in WK1 views on iPad.

* page/FrameView.cpp:
(WebCore::FrameView::viewportSizeForCSSViewportUnits):

LayoutTests:

New test that ensures that viewport units are resolved against the correct
viewport size after the first style recalc.

* fast/css/viewport-units-dynamic.html: Added.
* platform/mac/fast/css/viewport-units-dynamic-expected.txt: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@171567 268f45cc-cd09-0410-ab3c-d52691b4dbfc