Do not try to hit test a null mainFrameRenderView on dynamicViewportSizeUpdate() https://bugs.webkit.org/show_bug.cgi?id=135277 <rdar://problem/17804891>
Patch by Benjamin Poulain <bpoulain@apple.com> on 2014-07-24
Reviewed by Tim Horton.

* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::dynamicViewportSizeUpdate):
There is no guarantee that the main frame have its root view when performing a dynamicViewportSizeUpdate(),
we should not attempt to use the layer without null checking it first.

The odd part for me is <rdar://problem/17804891> is a little too frequent. In the vast majority of cases,
there is a RenderView, it seems actually pretty hard not to have one on dynamicViewportSizeUpdate().

Skipping hit testing is safe because it is a completely optional part of this algorithm.
When the hit test is not done, the new position is computed based on the relative position prior to
the size change.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@171576 268f45cc-cd09-0410-ab3c-d52691b4dbfc