Web Inspector: Incorrectly sized TimelineDataGrid event bubble https://bugs.webkit.org/show_bug.cgi?id=135371
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2014-07-28
Reviewed by Timothy Hatcher.

Previously the secondsPerPixel calculation was relying on an element
that may not be sized yet. This was resulting in a visibleWidth of 0
and resulted in secondsPerPixel being Infinity. Fortunately, the
graph data source already knows the secondsPerPixel so we can just
ask it. Getting the correct value and eliminating forced layouts.

* UserInterface/Views/OverviewTimelineView.js:
(WebInspector.OverviewTimelineView.prototype.get secondsPerPixel):
* UserInterface/Views/TimelineDataGridNode.js:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@171721 268f45cc-cd09-0410-ab3c-d52691b4dbfc