Assertion failure hit testing a region-based multicolumn in a region https://bugs.webkit.org/show_bug.cgi?id=135385
Reviewed by Andrei Bucur.

Source/WebCore:
When a region-based multicolumn element is displayed and hit tested in a region,
we have to disable the named flow region information not only for painting,
but also for hit-testing. This is a follow-up for https://bugs.webkit.org/show_bug.cgi?id=132121,
which provided the fix for painting.

Test: fast/regions/assert-hit-test-multicol-in-region.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::hitTestLayer):

LayoutTests:
* fast/regions/assert-hit-test-multicol-in-region-expected.txt: Added.
* fast/regions/assert-hit-test-multicol-in-region.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@171745 268f45cc-cd09-0410-ab3c-d52691b4dbfc