r171362 accidentally increased the size of InlineCallFrame. <https://webkit.org/b/136141>
Reviewed by Filip Pizlo.

r171362 increased the size of InlineCallFrame::kind to 2 bits.  This increased
the size of InlineCallFrame from 72 to 80 though not intentionally.  The fix
is to reduce the size of InlineCallFrame::stackOffset to 29 bits.

Also added an assert to ensure that we never set a value that exceeds the size
of InlineCallFrame::stackOffset.

* bytecode/CodeOrigin.h:
(JSC::InlineCallFrame::setStackOffset):
* dfg/DFGByteCodeParser.cpp:
(JSC::DFG::ByteCodeParser::InlineStackEntry::InlineStackEntry):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@172853 268f45cc-cd09-0410-ab3c-d52691b4dbfc