Rubber-banding in overflow:scroll regions does not work correctly with direction:rtl https://bugs.webkit.org/show_bug.cgi?id=136176
Reviewed by Sam Weinig.

Use scrollPosition() instead of scrollXOffset() and scrollYOffset() which both add
in the scrollOrigin(). Other implementations of visibleContentRectInternal() do
not add in the scrollOrigin(), so RenderLayer's shouldn't either in order to be
consistent. This makes rubber-banding work for RTL because
ScrollAnimatorMac::pinnedInDirection() and
ScrollAnimatorMac::absoluteScrollPosition() both return the expect values now.

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::visibleContentRectInternal):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@172945 268f45cc-cd09-0410-ab3c-d52691b4dbfc