Update placeholder-shown-basics.html for GTK
GTK does not support the color input type. This was causing placeholder-shown-basics.html
to fail when testing the non-support of placeholder by input[type=color].

Since this is a minor part of the test, it is more valuable getting the test running everywhere
than skipping it. I removed the subtest causing troubles.

Patch by Benjamin Poulain <bpoulain@apple.com> on 2014-08-27

* fast/css/placeholder-shown-basics-expected.html:
* fast/css/placeholder-shown-basics.html:
* platform/gtk/TestExpectations:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@173042 268f45cc-cd09-0410-ab3c-d52691b4dbfc