textStylingAtPosition returns incorrect values after executing toggleBold, toggleItalic and toggleUnderline. https://bugs.webkit.org/show_bug.cgi?id=136323 rdar://problem/18141964
Reviewed by Antti Koivisto.

For underline style we need to check typingStyle first and use that information to populate
the dictionary. If there is no typing style we can use the render style.
Source/WebCore:

* WebCore.exp.in:
* editing/ios/EditorIOS.mm:
(WebCore::Editor::fontAttributesForSelectionStart):

Source/WebKit2:

We also need to update the editor state for the toggle commands to reflect the state in the UIProcess
even for commands that don't change the selection.

* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::editorState):
* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::executeEditCommandWithCallback):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@173071 268f45cc-cd09-0410-ab3c-d52691b4dbfc