Configure Xcode to build a command line tool for the iOS Simulator https://bugs.webkit.org/show_bug.cgi?id=136388
Reviewed by David Kilzer.

Add a script, called configure-xcode-for-ios-development, to configure Xcode to
support building a command line tool for the iOS Simulator. You must run this
script as root.

The script uses the xcspec files in the OS X SDK to create xcspec files in the
iOS Simulator SDK with the product- and package- type definitions to build a
command line tool, if applicable.

* Scripts/configure-xcode-for-ios-development: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@173129 268f45cc-cd09-0410-ab3c-d52691b4dbfc