Bad cast in isValidColumnSpanner. https://bugs.webkit.org/show_bug.cgi?id=133380.
Reviewed by Simon Fraser.

Source/WebCore:

Added fast/multicol/multicol-crazy-nesting.html

* rendering/RenderMultiColumnFlowThread.cpp:
(WebCore::isValidColumnSpanner):
The crawl up the chain looking for unsplittable objects should use containing blocks
and not be casting to parent boxes. You can have an inline in the parent chain, so
this was not the right way to walk up the tree.

LayoutTests:

* fast/multicol/multicol-crazy-nesting-expected.txt: Added.
* fast/multicol/multicol-crazy-nesting.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@173845 268f45cc-cd09-0410-ab3c-d52691b4dbfc