REGRESSION (r165896): UIWebView PDFs have a forced black background https://bugs.webkit.org/show_bug.cgi?id=137387 <rdar://problem/17948411>
Reviewed by David Kilzer.

* WebView/WebPDFViewPlaceholder.h:
* WebView/WebPDFViewPlaceholder.mm:
(-[WebPDFViewPlaceholder layout]):
Remove an ill-advised attempt to force the PDF background color from WebKit.
The idea was that we would maintain Safari's black background except if there
were no pages in the PDF, in which case we would use white instead to avoid
having an all-black view. However, this broke clients who didn't want a black
background, because we were now forcing it to black. Also, we don't need to
maintain the black color in Safari because it doesn't use WebPDFViewPlaceholder anymore.
This wasn't an important part of the patch, in any case.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@174284 268f45cc-cd09-0410-ab3c-d52691b4dbfc