Highlight clicked links in WKPDFView https://bugs.webkit.org/show_bug.cgi?id=137400
Reviewed by Tim Horton.

Show a _UIHighlightView on top of clicked links for 200 ms before starting the navigation. This matches the behavior of UIWebPDFView.

* UIProcess/ios/WKPDFView.mm:
(-[WKPDFView _createHighlightViewWithFrame:]): Created a _UIHighlightView with a color and border radius matching the values used by UIWebPDFView.
(-[WKPDFView annotation:wasTouchedAtPoint:controller:]): Displayed the highlight, then removed it after the navigation began.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@174286 268f45cc-cd09-0410-ab3c-d52691b4dbfc