Fix remaining misuses of abs() and fabsf() https://bugs.webkit.org/show_bug.cgi?id=130947
Reviewed by Simon Fraser.

Fixed issues found when compiling with -Wabsolute-value.

* platform/ios/ScrollAnimatorIOS.mm:
(WebCore::ScrollAnimatorIOS::handleTouchEvent): Called abs() instead of fabsf for integral types. Explicitly
cast the divisor to float in the call to atanf().
* platform/ios/wak/WKView.mm:
(_WKViewAutoresizeCoord): Explicitly cast origMarginsTotal to an int before calling abs(). Added a FIXME
indicating the strange behavior when origMarginsTotal is in the range (0, 1) (as pointed out by Darin).

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@174383 268f45cc-cd09-0410-ab3c-d52691b4dbfc