Unify logic in RunTests.run() to run the Python unit tests https://bugs.webkit.org/show_bug.cgi?id=138160
Reviewed by Martin Robinson.

Unify the interactive and non-interactive code path for running the Python unit tests
in step RunTests so as to remove duplicate code to compute the the path to the script
test-webkitpy and log an informative message ("Running Python unit tests") to the console.

No new tests since no functionality was changed.

* Scripts/webkitpy/port/base.py:
(Port.python_unittest_results_directory):
* Scripts/webkitpy/tool/steps/runtests.py:
(RunTests.run):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@175295 268f45cc-cd09-0410-ab3c-d52691b4dbfc