Avoid synchronous style recalc in Document.activeElement <https://webkit.org/b/138850>
Reviewed by Ryosuke Niwa.

Source/WebCore:

This subtly changes the behavior of <input autofocus>, which was caught by
a layout test. Auto-focus doesn't happen until the render tree is attached,
although it was possible to force it by accessing document.activeElement.

Now auto-focus is always lazy. This is consistent with other browser engines.

~1% progression on Speedometer locally.

* dom/Document.cpp:
(WebCore::Document::activeElement):

LayoutTests:

Tweak a test to run once the DOM is loaded instead of during parsing.

* fast/forms/autofocus-in-sandbox-with-allow-scripts.html:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@176294 268f45cc-cd09-0410-ab3c-d52691b4dbfc