Add the data layout to the module and fix the pass order. https://bugs.webkit.org/show_bug.cgi?id=138748
Patch by Juergen Ributzka <juergen@apple.com> on 2014-12-05
Reviewed by Oliver Hunt.

This adds the data layout to the module, so it can be used by all
optimization passes in the LLVM optimizer pipeline. This also allows
FastISel to select more instructions, because less non-legal types are
generated.

Also fix the order of the alias analysis passes in the optimization
pipeline.

* ftl/FTLCompile.cpp:
(JSC::FTL::mmAllocateDataSection):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@176890 268f45cc-cd09-0410-ab3c-d52691b4dbfc