REGRESSION(r155906): Page content disappears on Tuaw article after loading https://bugs.webkit.org/show_bug.cgi?id=138100
Reviewed by Simon Fraser.

.:

DRT causes an extra paint which makes it impossible to test this with
an automated test.

* ManualTests/float-layer-not-painting.html: Added.

Source/WebCore:

Test: ManualTests/float-layer-not-painting.html

* rendering/RenderBlockFlow.cpp:
(WebCore::RenderBlockFlow::insertFloatingObject): If a layout happens,
it can change whether the float's renderer has a self painting layer.
So in that case, we need to update the flag on the FloatingObject
instance for that float.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@177049 268f45cc-cd09-0410-ab3c-d52691b4dbfc