Continuously repainting large parts of Huffington Post. https://bugs.webkit.org/show_bug.cgi?id=139468
Reviewed by Antti Koivisto.

This patch eliminates redundant repaint requests for inlines when neither the parent
block flow nor any of the inline children are dirty.
Previously,
1. simple line layout always recalculated inline content positions regardless of whether
the content needed relayout at all; as a result, it always triggered full repaint.
2. inline tree layout ignored the needslayout flag on the last line and treated it dirty
(unless it broke cleanly (<div>foo<br></div>)).
This was an ancient workaround for an editing/insert use case, but it seems not to be the case anymore.

Source/WebCore:

Tests: fast/repaint/implicitly-positioned-block-repaint-complex-line-layout.html
       fast/repaint/implicitly-positioned-block-repaint-simple-line-layout.html

* rendering/RenderBlockFlow.cpp:
(WebCore::RenderBlockFlow::layoutInlineChildren):
(WebCore::RenderBlockFlow::layoutSimpleLines): Check if we need to trigger layout at all.
* rendering/RenderBlockFlow.h:
* rendering/RenderBlockLineLayout.cpp:
(WebCore::RenderBlockFlow::determineStartPosition): Remove the last line dirty hack. If it
happens to introduce regression, we should fix it at the caller site to make the line dirty.

LayoutTests:

* fast/repaint/implicitly-positioned-block-repaint-complex-line-layout-expected.txt: Added.
* fast/repaint/implicitly-positioned-block-repaint-complex-line-layout.html: Added.
* fast/repaint/implicitly-positioned-block-repaint-simple-line-layout-expected.txt: Added.
* fast/repaint/implicitly-positioned-block-repaint-simple-line-layout.html: Added.
* platform/mac/fast/line-grid/line-align-right-edges-expected.txt:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@177128 268f45cc-cd09-0410-ab3c-d52691b4dbfc