Possible crash when setting the WKWebViews's scroll view delegate, if the scroll view outlives the web view https://bugs.webkit.org/show_bug.cgi?id=139662 rdar://problem/18083882
Reviewed by Benjamin Poulain.

Null out the internal delegate on the WKScrollView when the WKWebView goes away, since
it's possible for a client to set its own scroll view delegate, forcing the creation of
a WKScrollViewDelegateForwarder, and then retain the UIScrollView past the lifetime of
the WKWebView. In this situation, the WKScrollViewDelegateForwarder's internalDelegate
would point to a deleted WKWebView.

* UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView dealloc]):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@177329 268f45cc-cd09-0410-ab3c-d52691b4dbfc