Make explicit which TextIndicator animations are driven manually, and which run automatically https://bugs.webkit.org/show_bug.cgi?id=140113 <rdar://problem/19383425>
Reviewed by Darin Adler.

* UIProcess/mac/PageClientImpl.mm:
(WebKit::PageClientImpl::didPerformDictionaryLookup):
Delete an inaccurate comment.

* page/TextIndicator.cpp:
(WebCore::TextIndicator::wantsManualAnimation):
Add wantsManualAnimation(). The old transitions (Bounce and BounceAndCrossfade)
run automatically, and the new ones (FadeIn and Crossfade) are driven manually.

* page/TextIndicator.h:
Add a comment explaining which animations are manual and which are not.

* page/mac/TextIndicatorWindow.mm:
(-[WebTextIndicatorView present]):
Make use of wantsManualAnimation(). The previous (wantsFadeIn || wantsCrossfade)
was wrong, because it was true for BounceAndCrossfade, and would cause
BounceAndCrossfade animations (which aren't driven manually) to stall at progress=0.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@177970 268f45cc-cd09-0410-ab3c-d52691b4dbfc