Fix API tests that are broken under HiDPI https://bugs.webkit.org/show_bug.cgi?id=140146
Reviewed by Tim Horton.

* TestWebKitAPI/Tests/WebKit2/FindMatches.mm:
(TestWebKitAPI::didGetImageForMatchResult):
NSImages are measured in points, so grab the underlying bitmap image and get its size in pixels.

* TestWebKitAPI/Tests/mac/DeviceScaleFactorOnBack.mm:
(TestWebKitAPI::DeviceScaleFactorOnBack::runTest):
Don't assume that a web view not in a window has a backing scale factor of 1. Instead, always have web views start
out in a window that has a backing scale factor of 1.

* TestWebKitAPI/Tests/mac/DynamicDeviceScaleFactor.mm:
(TestWebKitAPI::DynamicDeviceScaleFactor::runTest):
Ditto.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@177997 268f45cc-cd09-0410-ab3c-d52691b4dbfc