Only throttle DOM timers if they change CSS properties that cannot cause non-descendant elements to become visible https://bugs.webkit.org/show_bug.cgi?id=140142
Reviewed by Darin Adler.

Source/WebCore:

Only throttle DOM timers if they change CSS properties that cannot cause
non-descendant elements to become visible. For example, a DOM timer may
change the height of an element that is outside the viewport. However,
changing its height may causes elements inside the viewport to move. As
such, we should NOT throttle such DOM timer.

With this patch, we only mark a repeating timer as candidate for
throttling if it is updating one of the following CSS properties:
left, right, top, bottom, webkit-transform, opacity.

Test: fast/dom/timer-height-change-no-throttling.html

* bindings/js/JSCSSStyleDeclarationCustom.cpp:
(WebCore::propertyChangeMayRepaintNonDescendants):
(WebCore::JSCSSStyleDeclaration::putDelegate):
* page/DOMTimer.cpp:
(WebCore::DOMTimer::scriptDidCauseElementRepaint):
* page/DOMTimer.h:

LayoutTests:

Add a layout test to make sure that a DOM timer changing the 'height'
property of an element outside the viewport does not get throttled.

* fast/dom/timer-height-change-no-throttling-expected.txt: Added.
* fast/dom/timer-height-change-no-throttling.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@178014 268f45cc-cd09-0410-ab3c-d52691b4dbfc