ASSERTION FAILED: rareData->m_flowThreadContainingBlock.value() == RenderBox::locateFlowThreadContainingBlock() in WebCore::RenderBlock::locateFlowThreadContainingBlock https://bugs.webkit.org/show_bug.cgi?id=139390
Reviewed by Andrei Bucur.

Source/WebCore:

When a block level element is inserted into an inline element, the inline element is split
in order to make sure that all the inline element children are properly wrapped inside
anonymous block elements. If the inline element is inside a pagination context, we have
to invalidate the cached flow thread containing block for the split inline element and its descendants.

Test: fast/multicol/newmulticol/block-split-inline-in-multicol.html

* rendering/RenderInline.cpp:
(WebCore::RenderInline::splitInlines):

LayoutTests:

* fast/multicol/newmulticol/block-split-inline-in-multicol-expected.txt: Added.
* fast/multicol/newmulticol/block-split-inline-in-multicol.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@178025 268f45cc-cd09-0410-ab3c-d52691b4dbfc