Secondary WebKit processes get re-exec'ed twice when running under ASan https://bugs.webkit.org/show_bug.cgi?id=140175
Reviewed by Sam Weinig.

* UIProcess/Launcher/mac/ProcessLauncherMac.mm: (WebKit::addDYLDEnvironmentAdditions):
If the main process is ASanified, add the ASan dynamic library to DYLD_INSERT_LIBRARIES.
This avoids an extra re-exec, which is very annoying when debugging.
The library is linked in using @rpath, so the only reliable way to find its location
is by examining binary images loaded into the current process.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@178043 268f45cc-cd09-0410-ab3c-d52691b4dbfc