Negative raster shape height leads to crash https://bugs.webkit.org/show_bug.cgi?id=136087
Reviewed by Zoltan Horvath.

Source/WebCore:

This patch clamps the size of the margin box rect used to compute the
raster shape on a float so the width and height cannot be negative.
This makes it safe to use that rect to determine the size of the shape
intervals data structure.

Test: fast/shapes/shape-outside-floats/shape-outside-negative-height-crash.html

* rendering/shapes/Shape.cpp:
(WebCore::Shape::createRasterShape): Add an assert to be explict about
    the precondition.
* rendering/shapes/ShapeOutsideInfo.cpp:
(WebCore::getShapeImageMarginRect): Clamp the marginRect's size.

LayoutTests:

* fast/shapes/shape-outside-floats/shape-outside-negative-height-crash-expected.txt: Added.
* fast/shapes/shape-outside-floats/shape-outside-negative-height-crash.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@178054 268f45cc-cd09-0410-ab3c-d52691b4dbfc