Update css3/flexbox/auto-margins.html after r177774 https://bugs.webkit.org/show_bug.cgi?id=139975
Patch by Myles C. Maxfield <litherum@gmail.com> on 2015-01-07
Reviewed by Darin Adler.

This is a subpixel problem. Flexbox uses SimpleLineLayout for the entire
box and therefore does all its math in floating point, while the
-expected uses the regular renderer hierarchy and therefore fixed-point
math. This patch updates the test to use Ahem so that the character
widths divide the same way in both cases.

* css3/flexbox/auto-margins-expected.html:
* css3/flexbox/auto-margins.html:
* platform/mac/TestExpectations:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@178063 268f45cc-cd09-0410-ab3c-d52691b4dbfc