Filters aren't applied to elements in columns after the first https://bugs.webkit.org/show_bug.cgi?id=140331 Source/WebCore:
Reviewed by Simon Fraser.

The important bits of this change came from Simon.

Filters and clipping were not taking columns into
account when using their offset rectangles. The fix
is to recalculate the rects if you're in such a
situation.

Tests: fast/multicol/clip-in-columns.html
       fast/multicol/filter-in-columns.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::hasFilterThatIsPainting): New method used
to check if we're about to paint a filter.
(WebCore::RenderLayer::setupFilters): Use the new helper if possible.
(WebCore::RenderLayer::paintLayerContents): If we're in columns,
and we either have a clip or a filter, recalculate the offset rectangles.
* rendering/RenderLayer.h:

LayoutTests:

<rdar://problem/19313858>

Reviewed by Simon Fraser.

Tests for filters and clipping on a column other
than the first.

* fast/multicol/clip-in-columns-expected.html: Added.
* fast/multicol/clip-in-columns.html: Added.
* fast/multicol/filter-in-columns-expected.html: Added.
* fast/multicol/filter-in-columns.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@178380 268f45cc-cd09-0410-ab3c-d52691b4dbfc