SVGNames.cpp and HTMLNames.cpp are sometimes malformed https://bugs.webkit.org/show_bug.cgi?id=140784 <rdar://problem/17701233>
Reviewed by David Kilzer.

Due to bad rules in DerivedSources.make, make_names.pl was called
several times with the same arguments during build. This is because
SVG/HTML/MathML each had several targets associated to the same
command. This would cause the command (make_names.pl) to be called
once per target. However, make_names.pl only needs to be executed
once to generate all of the files currently marked as target.

Calling make_names.pl several times for the same files is not only
wasteful, but it can also lead to malformed generated files if the
scripts instances are executed in parallel as they will read / write
from / to the same files.

This patch uses INTERMEDIATE targets instead so that we can still
indicate which files are generated by make_names.pl but only run
the script once.

This is a similar fix as for Bug 140756. There may several other
similar issues in DerivedSources.make but this patch focuses on
make_names.pl. I will take care of other in follow-ups if needed.

* DerivedSources.make:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@178952 268f45cc-cd09-0410-ab3c-d52691b4dbfc