Remove the various aliases for --ios-simulator (--sim, --simulator, --ios-sim) in arguments for build and test scripts https://bugs.webkit.org/show_bug.cgi?id=139748
Patch by Simon Fraser <simon.fraser@apple.com> on 2015-01-26
Reviewed by Daniel Bates.

There was much confusion resulting from the multitude of aliases used to build/test for
the iOS simulator. Standardize on "--ios-simulator", which is also the name of the layout
tests directory.

This fixes an issue where run-api-tests and some other scripts failed when run with
--simulator.

* Scripts/build-webkit:
* Scripts/package-root:
(usage):
* Scripts/webkitdirs.pm:
(argumentsForConfiguration):
(determineXcodeSDK):
* Scripts/webkitpy/port/factory.py:
(platform_options):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@179162 268f45cc-cd09-0410-ab3c-d52691b4dbfc