Remove the USE(ACCELERATED_COMPOSITING) guard from the MediaPlayerPrivateGStreamerBase destructor.
The guard was removed in r163079. It was still defined through CMake
for the EFL and GTK ports when the guarded code was introduced in
r172828. In r171741 the macro was finally removed for the GTK and EFL
ports as well, but the guards for the affected code weren't updated
so TextureMapperPlatformLayer::Client::platformLayerWillBeDestroyed()
wasn't being called anymore.

Rubber-stamped by Philippe Normand.

* platform/graphics/gstreamer/MediaPlayerPrivateGStreamerBase.cpp:
(WebCore::MediaPlayerPrivateGStreamerBase::~MediaPlayerPrivateGStreamerBase):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@179985 268f45cc-cd09-0410-ab3c-d52691b4dbfc