http/tests/xmlhttprequest/event-listener-gc.html is flaky https://bugs.webkit.org/show_bug.cgi?id=33342
Reviewed by Darin Adler.

Instead of using a zero delay timer, which can be legitimately delayed for dozens
of milliseconds, perform the collection in a second load event handler. GC code
has changed so much that I cannot confirm that the test still verifies the fix,
however we collect after returning for the function and unwinding the stack, so it should.

Also, clear the JS variable for a good measure.

* http/tests/xmlhttprequest/event-listener-gc.html:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@180105 268f45cc-cd09-0410-ab3c-d52691b4dbfc