REGRESSION(r179347): Clearing the PageCache no longer clears the PageCache. <https://webkit.org/b/141788>
Reviewed by Anders Carlsson.

Source/WebCore:

Once again we've fallen into the TemporaryChange trap:

    TemporaryChange<unsigned>(m_member, temporaryValue);

The code above doesn't actually do anything. Since the TemporaryChange local is not named,
it immediately goes out of scope and restores the original value of m_member.

Unless someone knows a C++ trick to prevent these, we'll need to add a style checker pass
to catch bugs like this. Whatever we do will be done separately from this bug.

Test: fast/history/page-cache-clearing.html

* history/PageCache.cpp:
(WebCore::PageCache::pruneToSizeNow): Name the local so it lives longer.

* testing/Internals.cpp:
(WebCore::Internals::clearPageCache):
(WebCore::Internals::pageCacheSize):
* testing/Internals.h:
* testing/Internals.idl: Add a way to clear the page cache and query its size from
window.internals to facilitate writing a simple test for this bug.

LayoutTests:

Add a simple test that navigates to a temporary page which immediately does a history.back
navigation. Upon returning to the first page, check that the page cache now has 1 entry,
and that clearing the page cache makes that entry go away.

* fast/history/page-cache-clearing-expected.txt: Added.
* fast/history/page-cache-clearing.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@180337 268f45cc-cd09-0410-ab3c-d52691b4dbfc