REGRESSION (r180018 ): Holding a rubber-band in place can get stuck https://bugs.webkit.org/show_bug.cgi?id=142020 -and corresponding- rdar://problem/19945216
Reviewed by Tom Horton.

It was a mistaken assumption that it was necessary to return false in the zero-
delta case. That is clearly conceptually wrong since false represents the DOM
doing something special with the event, which is clearly not the case if we never
even send the event to the DOM. Returning true will allow the rest of the
scrolling machinery the ability to handle the event.
* dom/Element.cpp:
(WebCore::Element::dispatchWheelEvent):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@180634 268f45cc-cd09-0410-ab3c-d52691b4dbfc