BuiltinExecutables keeps finalized Weaks around, pinning WeakBlocks. <https://webkit.org/b/142460>
Reviewed by Geoffrey Garen.

Make BuiltinExecutables a WeakHandleOwner so it can clear out its respective Weak members
if and when their pointees get garbage collected.

This fixes an issue I've seen locally where a WeakBlock is pinned down by a single one of
these Weak<BuiltinExecutables>.

* builtins/BuiltinExecutables.h: Make BuiltinExecutables inherit from WeakHandleOwner.

* builtins/BuiltinExecutables.cpp:
(JSC::BuiltinExecutables::finalize): Clear out the relevant member pointer when it's been
garbage collected. We use the WeakImpl's "context" field to pass the address of the member.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@181248 268f45cc-cd09-0410-ab3c-d52691b4dbfc