SyntaxChecker assertion is trapped with computed property name and getter https://bugs.webkit.org/show_bug.cgi?id=142863
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2015-03-20
Reviewed by Ryosuke Niwa.

Source/JavaScriptCore:

* parser/SyntaxChecker.h:
(JSC::SyntaxChecker::getName):
Remove invalid assert. Computed properties will not have a name
and the calling code is checking for null expecting it. The
AST path (non-CheckingPath) already does this without the assert
so it is well tested.

LayoutTests:

* js/basic-computed-property-name-expected.txt:
* js/script-tests/basic-computed-property-name.js:
(a.string_appeared_here.runTest):
Add tests that wrap the expression in a function, this
will test the syntax checker.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@181807 268f45cc-cd09-0410-ab3c-d52691b4dbfc