Perf dashboard should show relative change in values https://bugs.webkit.org/show_bug.cgi?id=143252
Reviewed by Antti Koivisto.

When a range of values are selected, show the percentage difference between the start and the end
in addition to the absolute value difference. When a single point is selected, show the relative
difference with respect to the previous point. Use two significant figures and always show plus sign
when the difference is positive.

* public/v2/app.js: Compute and format the relative difference.
* public/v2/chart-pane.css: Don't let commits view shrink itself when they're all collapsed.
* public/v2/index.html: Show the relative difference.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@182185 268f45cc-cd09-0410-ab3c-d52691b4dbfc