fast/dom/Window/timer-resume-on-navigation-back.html is flaky https://bugs.webkit.org/show_bug.cgi?id=33685
Reviewed by Brady Eidson.

If the timer fired before navigation, the test passed without going down the code
path that it was supposed to test. If it fired when navigation was in provisional
state, then dumping results was delayed until the navigation was finished, and
thus the intermediate page got dumped, and the test failed.

* fast/dom/Window/timer-resume-on-navigation-back.html:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@182838 268f45cc-cd09-0410-ab3c-d52691b4dbfc