Simple line layout: Wrong text offsetting when range does not start from the first renderer. https://bugs.webkit.org/show_bug.cgi?id=144167 rdar://problem/20639857
Reviewed by Simon Fraser.

This patch ensures that TextIterator returns the right text when the input range starts
from a sibling node.

TextIterator::m_previousTextLengthInFlow keeps track of the current node offset from the parent.
Source/WebCore:

It is required to map simple line layout runs to RenderText positions.
This patch sets the offset value when the iteration start with a sibling node.

Test: fast/text/range-text-with-simple-line-layout.html

* editing/TextIterator.cpp:
(WebCore::TextIterator::TextIterator):
(WebCore::TextIterator::handleTextNode):

LayoutTests:

it is required to map simple line layout runs to RenderText positions.
This patch sets the offset value when the iteration start with a sibling node.

* fast/text/range-text-with-simple-line-layout-expected.txt: Added.
* fast/text/range-text-with-simple-line-layout.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@183413 268f45cc-cd09-0410-ab3c-d52691b4dbfc