Asynchronous (or timed-out synchronous) resize flashes white instead of page background color https://bugs.webkit.org/show_bug.cgi?id=144468
Reviewed by Simon Fraser and Dan Bernstein.

* UIProcess/API/mac/WKView.mm:
(-[WKView _viewBackgroundColor]):
(-[WKView updateLayer]):
Use the pageExtendedBackgroundColor for the background color of the WKView,
if we have one (and aren't using a transparent background). This means
during asynchronous resize or other cases where the Web process isn't keeping
up with the UI process during a resize, we'll see background color instead of white.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@183647 268f45cc-cd09-0410-ab3c-d52691b4dbfc